jimt/CT-822 - refactored out useHandleDragEnd hook
