+++ apps/src/codebridge/FileBrowser/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from './useFileUploader';
 export * from './useFileUploadErrorCallback';
+export * from './useHandleDragEnd';
 export * from './useHandleFileUpload';
 export * from './usePrompts';
